@@ -1,10 +1,12 @@
 module ApplicationHelper
-  def title
-    base_title = "Ruby on Rails Tutorial Sample App"
-        if @title.nil?
-          base_title
-        else
-          "#{base_title} | #{@title}"
-        end  
-  end 
+
+  # Returns the full title on a per-page basis.
+  def full_title(page_title)
+    base_title = "Big Island Gigs | Inventory System"
+    if page_title.empty?
+      base_title
+    else
+      "#{base_title} | #{page_title}"
+    end
+  end
 end
