@@ -2,55 +2,44 @@ require 'digest'
 
 class User < ActiveRecord::Base
   
-  attr_accessor :password
-  attr_accessible :name, :email, :password, :password_confirmation  
+  attr_accessible :name, :email, :password, :password_confirmation
+  has_secure_password
+  has_many :microposts, dependent: :destroy
+  has_many :relationships, foreign_key: "follower_id", dependent: :destroy
+  has_many :followed_users, through: :relationships, source: :followed
   
-  email_regex = /\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i
-  
-  validates :name,  :presence => true,
-                    :length => { :maximum => 50 }
-  validates :email, :presence => true,
-                    :format => { :with => email_regex },
-                    :uniqueness => { :case_sensitive => false }
-                    
-  # Automatically create the virtual attribute 'password_confirmation'.
-  validates :password, :presence     => true,
-                       :confirmation => true,
-                       :length       => { :within => 6..40 }
-                       
-  before_save :encrypt_password
-  
-  # Return true if the user's password matches the submitted password.
-    def has_password?(submitted_password)
-      # Compare encrypted_password with the encrypted version of
-      # submitted_password.
-      encrypted_password == encrypt(submitted_password)
-    end
-    
-    def self.authenticate(email, submitted_password)
-      user = find_by_email(email)
-      return nil  if user.nil?
-      return user if user.has_password?(submitted_password)
-    end
+
+  before_save { |user| user.email = email.downcase }
+  before_save :create_remember_token
+
+  validates :name,  presence: true, length: { maximum: 50 }
+  VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i
+  validates :email, presence: true, format: { with: VALID_EMAIL_REGEX },
+                    uniqueness: { case_sensitive: false }
+  validates :password, length: { minimum: 6 }
+  validates :password_confirmation, presence: true
+
+  def feed
+    Micropost.where("user_id = ?", id)
+  end
+
+  def following?(other_user)
+    relationships.find_by_followed_id(other_user.id)    
+  end
+
+  def follow!(other_user)
+    relationships.create!(followed_id: other_user.id)
+  end
+
+  def unfollow!(other_user)
+    relationships.find_by_followed_id(other_user.id).destroy
+  end
+
 
   private
 
-    def encrypt_password
-      self.salt = make_salt unless has_password?(password)
-      self.encrypted_password = encrypt(password) 
-    end
-
-    def encrypt(string) 
-      secure_hash("#{salt}--#{string}")
-    end
-    
-    def make_salt
-      secure_hash("#{Time.now.utc}--#{password}")
-    end
-    
-    def secure_hash(string)
-      Digest::SHA2.hexdigest(string)
-    end
-  
-  
+	  def create_remember_token
+	  	self.remember_token = SecureRandom.urlsafe_base64
+	  end
+	  
 end
