@@ -1,7 +1,7 @@
 # By using the symbol ':user', we get Factory Girl to simulate the User model.
 Factory.define :user do |user|
-  user.name                  "Michael Hartl"
-  user.email                 "mhartl@example.com"
-  user.password              "foobar"
-  user.password_confirmation "foobar"
+  user.name                  "Jana Kimmich"
+  user.email                 "jana@jana.com"
+  user.password              "kimmich"
+  user.password_confirmation "kimmich"
 end
